# libflux: use iovec-like array over zmsg