@chu11 2021-07-14
hmm there was one LGTM warning.  I decided to comment out code with a note.

```
+        /* On first access index == 0 && iovcnt > 0 guaranteed
+         * Re-add check on if code changes. */
+        /* if (index == iovcnt) { */
+        /*     errno = EPROTO; */
+        /*     return -1; */
+        /* } */
```