@garlick 2021-07-19
Poking at this...

I'm seeing a throughput bump when I realloc 3 iovecs at a time (see bbc998c6e13cc805467974254486f0172aac86c5).  I'm not sure I`m testing the same way as you (and running 3 times may not be sufficient):
```
test:
  for i in $(seq 1 3) ; do src/test/throughput.py -n 2000 -x /bin/true; done

Commit 252ea49 (this PR):
  number of jobs: 2000
  submit time:    0.515 s (3883.3 job/s)
  script runtime: 42.535s
  job runtime:    42.429s
  throughput:     47.1 job/s (script:  47.0 job/s)
  number of jobs: 2000
  submit time:    0.531 s (3764.8 job/s)
  script runtime: 48.775s
  job runtime:    48.608s
  throughput:     41.1 job/s (script:  41.0 job/s)
  number of jobs: 2000
  submit time:    0.573 s (3490.4 job/s)
  script runtime: 52.081s
  job runtime:    51.942s
  throughput:     38.5 job/s (script:  38.4 job/s)

WIth batch iov in bbc998c:
  number of jobs: 2000
  submit time:    0.523 s (3827.0 job/s)
  script runtime: 38.429s
  job runtime:    38.313s
  throughput:     52.2 job/s (script:  52.0 job/s)
  number of jobs: 2000
  submit time:    0.549 s (3642.9 job/s)
  script runtime: 48.772s
  job runtime:    48.625s
  throughput:     41.1 job/s (script:  41.0 job/s)
  number of jobs: 2000
  submit time:    0.556 s (3598.8 job/s)
  script runtime: 48.032s
  job runtime:    47.906s
  throughput:     41.7 job/s (script:  41.6 job/s)
```
My thought was that wasting a view iov slots wouldn't be very costly, but calling realloc potentially dozens of times when receiving a long-path message might be.