@chu11 2021-07-21
Re-pushed with fixes per above review.

I also no longer call `realloc` on every loop iteration.  I yoinked and then tweaked the @garlick patch mentioned above.

Not knowing the optimal number of `N` chunks to alloc each time you call `realloc`, I added a printf to `msg_to_iovec()` to output the number of frames in a message being sent.  I gathered that output and ran some stats.  The mean number of frames was 4.68, the mode 4, and the median 5.  The median of 5 was "close" to 4 though.

Granted this is a benchmark and not a real world case, I ended up picking N = 4 for the `realloc` chunks.  It performed the best in my throughput tests below, although it wasn't a huge / obvious win.  I'm not sure choosing N = 3, 4 or 5 matter much as just not calling `realloc` on every iteration is the real win.

realloc 1

```
Running throughput.py 2048 jobs per iteration, 5 iterations
throughput:     20.6 job/s (script:  20.6 job/s)
1: 100 seconds
throughput:     19.4 job/s (script:  19.3 job/s)
2: 206 seconds
throughput:     18.7 job/s (script:  18.6 job/s)
3: 316 seconds
throughput:     18.4 job/s (script:  18.4 job/s)
4: 428 seconds
throughput:     19.3 job/s (script:  19.2 job/s)
5: 535 seconds
Total of 535 seconds for everything
```

realloc 3

```
Running throughput.py 2048 jobs per iteration, 5 iterations
throughput:     20.6 job/s (script:  20.5 job/s)
1: 100 seconds
throughput:     19.4 job/s (script:  19.3 job/s)
2: 206 seconds
throughput:     19.3 job/s (script:  19.2 job/s)
3: 313 seconds
throughput:     20.1 job/s (script:  20.0 job/s)
4: 416 seconds
throughput:     20.2 job/s (script:  20.2 job/s)
5: 518 seconds
Total of 518 seconds for everything
```

realloc 4

```
Running throughput.py 2048 jobs per iteration, 5 iterations
throughput:     21.5 job/s (script:  21.4 job/s)
1: 96 seconds
throughput:     20.1 job/s (script:  20.0 job/s)
2: 198 seconds
throughput:     19.9 job/s (script:  19.9 job/s)
3: 302 seconds
throughput:     19.3 job/s (script:  19.2 job/s)
4: 408 seconds
throughput:     19.1 job/s (script:  19.0 job/s)
5: 516 seconds
Total of 516 seconds for everything
```

realloc 5

```
Running throughput.py 2048 jobs per iteration, 5 iterations
throughput:     20.8 job/s (script:  20.8 job/s)
1: 99 seconds
throughput:     19.5 job/s (script:  19.4 job/s)
2: 205 seconds
throughput:     19.5 job/s (script:  19.4 job/s)
3: 310 seconds
throughput:     19.7 job/s (script:  19.6 job/s)
4: 415 seconds
throughput:     19.1 job/s (script:  19.1 job/s)
5: 522 seconds
Total of 522 seconds for everything
```